Move fonts to config folder
